+++ test/helper.rb
@@ -23,6 +23,7 @@ end
 require 'minitest/autorun'
 require 'minitest/pride'
 
+require 'celluloid/current'
 require 'celluloid/test'
 Celluloid.boot
 require 'sidekiq'