+++ myapp/config/application.rb
@@ -55,5 +55,6 @@ module Myapp
 
     # Version of your assets, change this if you want to expire all your assets
     config.assets.version = '1.0'
+    config.secret_key_base = "sidekiq"
   end
 end