+++ test/test_sidekiq.rb
@@ -66,4 +66,22 @@ class TestSidekiq < Sidekiq::Test
       assert_equal 'cat', Sidekiq.default_worker_options['queue']
     end
   end
+
+  describe 'error handling' do
+    it 'deals with user-specified error handlers which raise errors' do
+      output = capture_logging do
+        begin
+          Sidekiq.error_handlers << proc {|x, hash|
+            raise 'boom'
+          }
+          cli = Sidekiq::CLI.new
+          cli.handle_exception(RuntimeError.new("hello"))
+        ensure
+          Sidekiq.error_handlers.pop
+        end
+      end
+      assert_includes output, "boom"
+      assert_includes output, "ERROR"
+    end
+  end
 end